Move domain directory
