+++ src/domain/entities/UserDetail.ts
@@ -1,7 +1,7 @@
 interface UserDetail {
   id?: number;
-  email: string;
   name: string;
+  email: string;
   createdAt: string;
   updatedAt: string;
 }
